+++ src/api/Migrations/20240723071757_MakeReliabilityIdNullable.cs
@@ -1,0 +1,27 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace BDMS.Migrations;
+
+/// <inheritdoc />
+public partial class MakeReliabilityIdNullable : Migration
+{
+    /// <inheritdoc />
+    protected override void Up(MigrationBuilder migrationBuilder)
+    {
+        migrationBuilder.AlterColumn<int>(
+            name: "reliability",
+            schema: "bdms",
+            table: "observation",
+            type: "integer",
+            nullable: true,
+            oldClrType: typeof(int),
+            oldType: "integer");
+    }
+
+    /// <inheritdoc />
+    protected override void Down(MigrationBuilder migrationBuilder)
+    {
+    }
+}